Update core-im-source.yaml (#65)
Proposing to change `subjectClassification` back to `classification`.  We apply the pattern of renaming the core-im `subject` property by appending the type of thing the subject is - e,g. `subjectVariant`.  It is therefore confusing to see an attribute called `subjectClassification`, as it implies that a Classification is the subject of the statement.